Support size limitations on channel queues
Given a maxsize, fill the queue until we reach the limit. Upon reaching
that limit, don't even read until the space becomes available. If a
command would overfill an almost-full queue, fit as much of that command
into the queue as possible and continue adding the rest when space
becomes available.